Refactor `MovieRow` and `MovieModal`
